% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmsheet.R
\name{makeNonmemSheet}
\alias{makeNonmemSheet}
\title{Create data frame for NONMEM}
\usage{
makeNonmemSheet(df, selected_bm, selected_cov, XMAX = NULL, lognorm = TRUE)
}
\arguments{
\item{df}{The input data frame."ID" and "TIME" columns should be prepared in advance.}

\item{selected_bm}{A character vector specifying the selected biomarkers.}

\item{selected_cov}{A character vector specifying the selected covariates.}

\item{XMAX}{A numeric vector specifying the maximum values for biomarkers. Default is NULL.}

\item{lognorm}{A logical value indicating whether to perform log-normalization. Default is TRUE.}
}
\value{
A data frame for NONMEM.
}
\description{
This function takes a data frame and selected biomarkers and covariates as inputs, and generates a NONMEM sheet.
}
\details{
The function filters the data frame, performs calculations on biomarkers, and creates the NONMEM sheet with calculated means, counts, and other relevant columns.
}
\examples{
df <- read.csv("data.csv")
selected_bm <- c("biomarker1", "biomarker2")
selected_cov <- c("covariate1", "covariate2")
sheet <- makeNonmemSheet(df, selected_bm, selected_cov)

}
