% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmsheet.R
\name{scoreConverter}
\alias{scoreConverter}
\title{Score Converter}
\usage{
scoreConverter(x_vec, max)
}
\arguments{
\item{x_vec}{A numeric vector of values to be converted.}

\item{max}{The maximum value for the conversion.}
}
\value{
A vector of converted scores.
}
\description{
This function takes a vector of values and a maximum value as inputs, and performs a score conversion.
}
\details{
The function applies a score conversion formula to each value in the input vector. It adds 0.5 to each value, divides it by (max + 1), and performs a transformation to convert it to a score.
}
\examples{
x <- c(1, 2, 3)
max_value <- 5
converted_scores <- scoreConverter(x, max_value)

}
