% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmsheet.R
\name{setInitialPrms}
\alias{setInitialPrms}
\title{Set Initial Parameters}
\usage{
setInitialPrms(
  df,
  selected_bm,
  definition_bm,
  definition_value,
  XMAX = NULL,
  estimated_mean_offsetT = 10,
  lognorm = TRUE
)
}
\arguments{
\item{df}{The data frame containing the biomarker data.}

\item{selected_bm}{A character vector specifying the selected biomarkers.}

\item{definition_bm}{The biomarker used for definition.}

\item{definition_value}{The value used for the definition of the biomarker.}

\item{XMAX}{An optional parameter specifying the maximum value for log-normal conversion.}

\item{estimated_mean_offsetT}{The estimated mean offset.}

\item{lognorm}{Logical value indicating whether log-normal conversion should be applied.}
}
\value{
A data frame containing the calculated initial parameters.
}
\description{
This function sets the initial parameters based on the given data frame and biomarker information.
}
\examples{
setInitialPrms(df, selected_bm, definition_bm, definition_value)

}
