% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sreft-package.R
\docType{package}
\name{sreft-package}
\alias{sreft}
\alias{sreft-package}
\title{sreft: Utility tools for SReFT}
\description{
This package contains functions to preprocess data for SReFT (Statistical Restoration of Fragmented Time course) analysis and to plot the data after the SReFT analysis.SReFT analysis itself requires NONMEM or Napp, which are not supported by this package.
}
\keyword{internal}
